+++ samples/custom-enricher/app/src/main/java/io/fabric8/maven/sample/enricher/app/HelloController.java
@@ -19,6 +19,9 @@ import org.springframework.util.FileCopyUtils;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.io.FileReader;
+import java.io.IOException;
+
 
 /**
  * @author roland