+++ CHANGELOG.md
@@ -16,7 +16,8 @@ After this we will switch probably to real [Semantic Versioning 2.0.0](http://se
 * Fix #1697: NullpointerException when trying to apply custom resources
 * Fix #1676: Support for latest kubernetes client
 * Feature #1536: Java Image Builder Support
-* Fix #1704: fabric8-build failing on openshift 
+* Fix #1704: fabric8-build failing on openshift
+* Feature #1706: Prometheus Enricher; Configuration support for Prometheus path
 
 ### 4.2.0 (01-08-2019)
 * Fix #1638: Remove enrichAll parameter from ImageChangeTriggerEnricher