+++ CHANGELOG.md
@@ -14,6 +14,7 @@ After this we will switch probably to real [Semantic Versioning 2.0.0](http://se
 * Updated custom-enricher sample
 * Fixed image config in thorntail sample
 * Fix #1697: NullpointerException when trying to apply custom resources
+* Fix #1696: fmp not setting imagestreams resourceVersion properly.
 * Fix #1676: Support for latest kubernetes client
 * Feature #1536: Java Image Builder Support
 * Fix #1704: fabric8-build failing on openshift