+++ CHANGELOG.md
@@ -13,6 +13,7 @@ After this we will switch probably to real [Semantic Versioning 2.0.0](http://se
 ### 4.3-SNAPSHOT
 * Updated custom-enricher sample
 * Fixed image config in thorntail sample
+* Fix Autotls enricher in order to add service serving certificate secrets annotations
 * Fix #1697: NullpointerException when trying to apply custom resources
 * Fix #1696: fmp not setting imagestreams resourceVersion properly.
 * Fix #1676: Support for latest kubernetes client