+++ CHANGELOG.md
@@ -15,6 +15,7 @@ After this we will switch probably to real [Semantic Versioning 2.0.0](http://se
 * Fixed image config in thorntail sample
 * Fix #1697: NullpointerException when trying to apply custom resources
 * Fix #1696: fmp not setting imagestreams resourceVersion properly.
+* Fix #1689: HELM mode does not support parameters without value
 * Fix #1676: Support for latest kubernetes client
 * Fix #1699: Ability to specify object namespace in fragments 
 * Feature #1536: Java Image Builder Support