+++ CHANGELOG.md
@@ -11,6 +11,8 @@ We use semantic versioning in some slight variation until our feature set has st
 After this we will switch probably to real [Semantic Versioning 2.0.0](http://semver.org/)
 
 ### 4.3-SNAPSHOT
+
+### 4.3.0 (04-10-2019)
 * Updated custom-enricher sample
 * Fixed image config in thorntail sample
 * Fix Autotls enricher in order to add service serving certificate secrets annotations